Extract period status orchestration from domain module
