+++ src/application/summary-logs/period-status.test.js
@@ -1,8 +1,7 @@
 import { describe, expect, it, vi } from 'vitest'
 import {
   buildTransactionAmounts,
-  classifyByPeriodStatus,
-  computeLoadsByPeriodStatus
+  classifyByPeriodStatus
 } from './period-status.js'
 import { ROW_OUTCOME } from '#domain/summary-logs/table-schemas/validation-pipeline.js'
 import { ORS_VALIDATION_DISABLED } from '#domain/summary-logs/table-schemas/shared/classification-reason.js'
@@ -12,8 +11,6 @@ import { VERSION_STATUS } from '#domain/waste-records/model.js'
 /** @import {PeriodicReport} from '#reports/repository/port.js' */
 /** @import {Registration} from '#domain/organisations/registration.js' */
 /** @import {WasteRecord} from '#domain/waste-records/model.js' */
-/** @import {ReportsRepository} from '#reports/repository/port.js' */
-/** @import {TypedLogger} from '#common/helpers/logging/logger.js' */
 /** @import {TableSchema} from '#domain/summary-logs/table-schemas/index.js' */
 
 const accreditedRegistration = /** @type {Registration} */ (
@@ -1172,207 +1169,5 @@ describe('buildTransactionAmounts', () => {
     })
 
     expect(result.get('received:1000')).toEqual({ oldAmount: 0, newAmount: 15 })
-  })
-})
-
-describe('computeLoadsByPeriodStatus', () => {
-  const stubLogger = /** @type {TypedLogger} */ (
-    /** @type {unknown} */ ({
-      warn: vi.fn()
-    })
-  )
-
-  const TABLE_SCHEMAS_FOR_PROCESSING =
-    /** @type {typeof import('#domain/summary-logs/table-schemas/index.js').PROCESSING_TYPE_TABLES} */ (
-      /** @type {unknown} */ ({
-        REPROCESSOR_INPUT: {
-          RECEIVED_LOADS_FOR_REPROCESSING: {
-            reportingDateFields: ['DATE_RECEIVED_FOR_REPROCESSING'],
-            wasteRecordType: 'received',
-            classifyForWasteBalance: () => ({
-              outcome: ROW_OUTCOME.INCLUDED,
-              reasons: [],
-              transactionAmount: 10
-            })
-          }
-        }
-      })
-    )
-
-  it('returns classified loads when reports lookup succeeds', async () => {
-    const wasteRecords = [buildWasteRecord()]
-
-    const result = await computeLoadsByPeriodStatus({
-      wasteRecords,
-      wasteBalanceRecords: wasteRecords,
-      summaryLogId: SUMMARY_LOG_ID,
-      status: 'validated',
-      registration: accreditedRegistration,
-      processingType: 'REPROCESSOR_INPUT',
-      existingRecordsMap: new Map(),
-      reportsRepository: /** @type {ReportsRepository} */ (
-        /** @type {unknown} */ ({
-          findPeriodicReports: async () => []
-        })
-      ),
-
-      organisationId: 'org-1',
-      registrationId: 'reg-1',
-      loggingContext: 'test',
-      logger: stubLogger,
-      processingTypeTables: TABLE_SCHEMAS_FOR_PROCESSING
-    })
-
-    expect(result).not.toBeNull()
-    expect(
-      /** @type {NonNullable<typeof result>} */ (result).open?.added?.included
-        .tonnageDelta
-    ).toBe(10)
-  })
-
-  it('returns null and logs a warning when reports lookup fails', async () => {
-    vi.mocked(stubLogger.warn).mockClear()
-
-    const result = await computeLoadsByPeriodStatus({
-      wasteRecords: [buildWasteRecord()],
-      wasteBalanceRecords: [buildWasteRecord()],
-      summaryLogId: SUMMARY_LOG_ID,
-      status: 'validated',
-      registration: accreditedRegistration,
-      processingType: 'REPROCESSOR_INPUT',
-      existingRecordsMap: new Map(),
-      reportsRepository: /** @type {ReportsRepository} */ (
-        /** @type {unknown} */ ({
-          findPeriodicReports: async () => {
-            throw new Error('database unavailable')
-          }
-        })
-      ),
-      organisationId: 'org-1',
-      registrationId: 'reg-1',
-      loggingContext: 'test-context',
-      logger: stubLogger,
-      processingTypeTables: TABLE_SCHEMAS_FOR_PROCESSING
-    })
-
-    expect(result).toBeNull()
-    expect(stubLogger.warn).toHaveBeenCalledWith(
-      expect.objectContaining({
-        message: expect.stringContaining('test-context')
-      })
-    )
-  })
-
-  it('handles unknown waste record type within a known processing type', async () => {
-    const wasteRecords = [buildWasteRecord({ wasteRecordType: 'unknown' })]
-
-    const result = await computeLoadsByPeriodStatus({
-      wasteRecords,
-      wasteBalanceRecords: wasteRecords,
-      summaryLogId: SUMMARY_LOG_ID,
-      status: 'validated',
-      registration: accreditedRegistration,
-      processingType: 'REPROCESSOR_INPUT',
-      existingRecordsMap: new Map(),
-      reportsRepository: /** @type {ReportsRepository} */ (
-        /** @type {unknown} */ ({
-          findPeriodicReports: async () => []
-        })
-      ),
-
-      organisationId: 'org-1',
-      registrationId: 'reg-1',
-      loggingContext: 'test',
-      logger: stubLogger,
-      processingTypeTables: TABLE_SCHEMAS_FOR_PROCESSING
-    })
-
-    expect(result).toEqual({
-      open: {
-        added: {
-          included: { count: 1, tonnageDelta: 0 },
-          excluded: { count: 0, tonnageDelta: 0 }
-        },
-        adjusted: emptyChange()
-      },
-      closed: emptyPeriod()
-    })
-  })
-
-  it('passes null accreditation for registered-only registrations', async () => {
-    const classifyForWasteBalance = vi.fn().mockReturnValue({
-      outcome: ROW_OUTCOME.INCLUDED,
-      reasons: [],
-      transactionAmount: 5
-    })
-
-    const schemas =
-      /** @type {typeof import('#domain/summary-logs/table-schemas/index.js').PROCESSING_TYPE_TABLES} */ (
-        /** @type {unknown} */ ({
-          REPROCESSOR_INPUT: {
-            RECEIVED_LOADS_FOR_REPROCESSING: {
-              reportingDateFields: ['DATE_RECEIVED_FOR_REPROCESSING'],
-              wasteRecordType: 'received',
-              classifyForWasteBalance
-            }
-          }
-        })
-      )
-
-    const wasteRecords = [buildWasteRecord()]
-
-    await computeLoadsByPeriodStatus({
-      wasteRecords,
-      wasteBalanceRecords: wasteRecords,
-      summaryLogId: SUMMARY_LOG_ID,
-      status: 'validated',
-      registration: registeredOnlyRegistration,
-      processingType: 'REPROCESSOR_INPUT',
-      existingRecordsMap: new Map(),
-      reportsRepository: /** @type {ReportsRepository} */ (
-        /** @type {unknown} */ ({
-          findPeriodicReports: async () => []
-        })
-      ),
-
-      organisationId: 'org-1',
-      registrationId: 'reg-1',
-      loggingContext: 'test',
-      logger: stubLogger,
-      processingTypeTables: schemas
-    })
-
-    expect(classifyForWasteBalance).toHaveBeenCalledWith(
-      expect.any(Object),
-      expect.objectContaining({ accreditation: null })
-    )
-  })
-
-  it('returns null for unknown processing type', async () => {
-    vi.mocked(stubLogger.warn).mockClear()
-    const wasteRecords = [buildWasteRecord()]
-
-    const result = await computeLoadsByPeriodStatus({
-      wasteRecords,
-      wasteBalanceRecords: wasteRecords,
-      summaryLogId: SUMMARY_LOG_ID,
-      status: 'validated',
-      registration: accreditedRegistration,
-      processingType: 'UNKNOWN_TYPE',
-      existingRecordsMap: new Map(),
-      reportsRepository: /** @type {ReportsRepository} */ (
-        /** @type {unknown} */ ({
-          findPeriodicReports: async () => []
-        })
-      ),
-
-      organisationId: 'org-1',
-      registrationId: 'reg-1',
-      loggingContext: 'test',
-      logger: stubLogger,
-      processingTypeTables: TABLE_SCHEMAS_FOR_PROCESSING
-    })
-
-    expect(result).toBeNull()
   })
 })
